requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# Vector Search
pinecone-client==3.0.2
openai==1.63.0
numpy==1.26.4

# Storage
boto3==1.34.34
botocore==1.34.34

# Workflow
temporalio==1.5.0

# Text Processing
spacy==3.7.2
